put base offscreen class in new module
