+++ wgpu/gui/__init__.py
@@ -3,4 +3,4 @@ Code to provide a canvas to render to.
 """
 
 from .base import WgpuCanvasInterface, WgpuCanvasBase  # noqa: F401
-from .offscreen import WgpuOffscreenCanvas  # noqa: F401
+from ._offscreen import WgpuOffscreenCanvas  # noqa: F401
